+++ docs/examples/textfield/dontGeneric.tsx
@@ -1,0 +1,20 @@
+import { useState } from 'react';
+import { Box, TextField } from 'gestalt';
+
+export default function Example() {
+  const [password, setPassword] = useState('');
+
+  return (
+    <Box padding={8} width="100%">
+      <TextField
+        autoComplete="new-password"
+        errorMessage="There is an error"
+        id="best-practices-dont-error-message"
+        label="Password"
+        onChange={({ value }) => setPassword(value)}
+        type="password"
+        value={password}
+      />
+    </Box>
+  );
+}
