--- docs/examples/textfield/ensureVisibleLabel.tsx
@@ -1,23 +1,0 @@
-import { useState } from 'react';
-import { Box, TextField } from 'gestalt';
-
-export default function Example() {
-  const [username, setUsername] = useState('');
-
-  return (
-    <Box padding={8} width="100%">
-      <Box color="light" padding={2}>
-        <TextField
-          autoComplete="username"
-          id="best-practices-do-label"
-          label="Username"
-          onChange={({ value }) => {
-            setUsername(value);
-          }}
-          type="text"
-          value={username}
-        />
-      </Box>
-    </Box>
-  );
-}
