+++ packages/gestalt/src/TextField/VRInternalTextField.css
@@ -1,6 +1,7 @@
 .input {
   border-radius: var(--sema-rounding-md);
   box-sizing: border-box;
+  color: var(--sema-color-text-default);
   display: block;
   font-weight: var(--sema-font-weight-body);
   line-height: 16px;
@@ -39,6 +40,7 @@
 }
 
 .label {
+  color: var(--sema-color-text-default);
   font-size: var(--sema-font-size-ui-xs);
   font-weight: var(--sema-font-weight-ui);
   line-height: var(--sema-font-lineheight-UI-XS);
